# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Verspätungsbegleiter API
  version: "1.0.0"
  license:
    name: MIT
    identifier: MIT
  description: |
    Routing and travel companion API for Deutsche Bahn journeys.

    **No authentication.** Abuse-shaped via `X-Install-Id` UUID header + IP rate limiting.
    API is open source — see repository for details.

    All timestamps are UTC ISO 8601. Frontend is responsible for local timezone formatting.

    Error responses use RFC 7807 `application/problem+json` with URN-based `type` identifiers.
    Error catalog: `urn:vbb:error:<slug>` — see error responses below.
servers:
  - url: /v1
    description: Current version
...
security: []
paths:
...
  /stations:
    get:
      operationId: searchStations
      summary: Station name autocomplete (proxied from db.transport.rest, Redis-cached 5min)
      tags: [Stations]
      parameters:
        - $ref: '#/components/parameters/InstallId'
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 2
          description: "Search query, minimum 2 characters. Prefix + substring match, diacritic-insensitive."
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Maximum number of results
      responses:
        "200":
          description: "Station results. Empty array when no match — never 404."
          headers:
            Cache-Control:
              schema:
                type: string
              example: "public, max-age=300"
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationList'
              example:
                stations:
                  - id: "8000105"
                    name: "Frankfurt (Main) Hbf"
                  - id: "8000104"
                    name: "Frankfurt (Main) Süd"
        "400":
          $ref: '#/components/responses/MalformedRequest'
        "429":
          $ref: '#/components/responses/TooManyRequests'
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
    InstallId:
      in: header
      name: X-Install-Id
      schema:
        type: string
        format: uuid
      description: |
        UUID v4 device identifier for abuse shaping.
        Generated on first app launch, stored in IndexedDB + localStorage backup.
        Falls back to IP-based rate limiting if absent.
        Not authentication — open API.
...
  headers:
    RateLimitLimit:
      schema:
        type: integer
      description: Request limit per window
      example: 60
    RateLimitRemaining:
      schema:
        type: integer
      description: Requests remaining in current window
      example: 47
    RateLimitReset:
      schema:
        type: integer
      description: Unix timestamp when window resets
      example: 1749600420
    RequestId:
      schema:
        type: string
        format: uuid
      description: Server-generated request ID for log correlation. Echoed in error instance field.
  schemas:
...
    StationRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          description: HAFAS station ID
          example: "8000105"
        name:
          type: string
          example: "Frankfurt (Main) Hbf"
    StationList:
      type: object
      required: [stations]
      properties:
        stations:
          type: array
          items:
            $ref: '#/components/schemas/StationRef'
          description: "Empty array when no match. Never 404 for missing results."
...
    Problem:
      type: object
      required: [type, title, status]
      description: "RFC 7807 Problem Details. type is a stable URN identifier, not a dereferenceable URL."
      properties:
        type:
          type: string
          description: "URN error identifier. Format: urn:vbb:error:<slug>"
          example: "urn:vbb:error:train-not-found"
        title:
          type: string
          example: "Train Not Found"
        status:
          type: integer
          example: 404
        detail:
          type: string
          description: Human-readable, actionable error description.
          example: "Train ICE 123 does not operate on 2026-06-10."
        instance:
          type: string
          description: "Request path + X-Request-Id for log correlation."
          example: "/v1/trains/ICE123"
        errors:
          type: array
          description: "Field-level validation errors. Present on 422 responses only."
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: "destination"
              message:
                type: string
                example: "Must be a valid HAFAS station ID."
  responses:
    MalformedRequest:
      description: "Malformed request — unparseable JSON or missing Content-Type"
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "urn:vbb:error:malformed-request"
            title: "Malformed Request"
            status: 400
            detail: "Request body is not valid JSON."
            instance: "/v1/journeys"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: "Seconds to wait. Client should apply exponential backoff: min(Retry-After × 2^n, 300)."
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "urn:vbb:error:rate-limit-exceeded"
            title: "Rate Limit Exceeded"
            status: 429
    ServiceUnavailable:
      description: "Backend dependency unavailable (db.transport.rest unreachable or MAX_ACTIVE_JOURNEYS capacity reached)"
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: "urn:vbb:error:upstream-unavailable"
            title: "Service Unavailable"
            status: 503
...